+++ app/controllers/posts.rb
@@ -1,0 +1,17 @@
+class Chitter < Sinatra::Base
+
+  get '/posts/all' do
+    @peeps = Peep.all
+    erb(:'posts/all')
+  end
+
+  post '/posts/new' do
+    peep = Peep.create(content: params[:new_peep], posted_at: Time.now, user_id: session[:user_id])
+    redirect('/posts/all')
+  end
+
+  get '/posts/new' do
+    erb(:'posts/new')
+  end
+
+end
