+++ app/controllers/sessions.rb
@@ -1,0 +1,24 @@
+class Chitter < Sinatra::Base
+
+  get '/sessions/new' do
+    erb :'sessions/new'
+  end
+
+  post '/sessions' do
+    user = User.authenticate(params[:email], params[:password])
+    if user
+      session[:user_id] = user.id
+      redirect to('/posts/all')
+    else
+      flash.now[:errors] = 'The email or password is incorrect'
+      erb :'sessions/new'
+    end
+  end
+
+  delete '/sessions' do
+    session[:user_id] = nil
+    flash.keep[:notice] = 'See you again!'
+    redirect to '/posts/all'
+  end
+
+end
